% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_to_args.R
\name{argsToFlagList}
\alias{argsToFlagList}
\title{Convert list of function arguments to list of command flags}
\usage{
argsToFlagList(args, flag_lookup = NULL)
}
\arguments{
\item{args}{named list output from get*Args family of functions.}

\item{flag_lookup}{optional named vector used to convert args to command flags}
}
\value{
named list
}
\description{
Function also handles error checking to ensure args contain valid data types,
and looks for common usage mistakes.
}
\details{
For simple shell wrappers, this function isn't needed. However, the list
structure is more amenable to manipulation by package developers for advanced
use before evaluating them to the command flags vector.
}
\examples{
theFunction <- function(...){getAllArgs()}
theArgs <- theFunction(arg1 = "value", arg2 = TRUE)
argsToFlagList(theArgs)
}
