% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{build_is_valid_install}
\alias{build_is_valid_install}
\title{Macro for constructing boolean check for valid path}
\usage{
build_is_valid_install(path_handler)
}
\arguments{
\item{path_handler}{function output of \code{build_path_handler()} \strong{NOTE:} When
passing the function, do not pass as: \code{fun()}, but \code{fun} to avoid evaluation.}
}
\value{
a function returning TRUE or FALSE if a valid install is detected.
}
\description{
Macro for constructing boolean check for valid path
}
\examples{
handle <- build_path_handler(option_name = "meme_bin", default_path = "~/meme/bin")
valid_install <- build_is_valid_install(handle)
# Returns TRUE is "~/meme/bin/" exists
valid_install()
# Returns FALSE if "bad/path/" doesn't exist
valid_install("bad/path/")

# Also works with options
handle_option_only <- build_path_handler(option_name = "meme_bin")
valid_install2 <- build_is_valid_install(handle_option_only)
options(meme_bin = "~/meme/bin/")
valid_install2()
}
