% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_args_contain_illegal_flags}
\alias{check_args_contain_illegal_flags}
\title{Checks dots for illegal flag names}
\usage{
check_args_contain_illegal_flags(dots)
}
\arguments{
\item{dots}{list output of getDots}
}
\value{
prints warning for each illegal flag
}
\description{
Checks dots for illegal flag names
}
\examples{
theFunction <- function(...) { getDotArgs() }
theDots <-  theFunction(validFlag = TRUE, `&illegalFlag` = "will error")
\dontrun{
theArgs <-  dotsToArgs(theDots)
}
}
