% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_install}
\alias{check_install}
\title{Wrapper function for checking an install}
\usage{
check_install(path_handler, path = NULL)
}
\arguments{
\item{path_handler}{\code{function} output of \code{build_path_handler()}}

\item{path}{user-override path to check (identical to \code{path} argument of \code{build_path_handler()} output)}
}
\value{
pretty printed message indicating whether files exits or not. Green check = Yes, red X = No.
}
\description{
This function can be lightly wrapped by package builders to build a user-friendly install checking function.
}
\examples{
\dontrun{

check_install()
}
}
