% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_outputs}
\alias{check_outputs}
\title{Creates list of paths by file extension & checks they exist}
\usage{
check_outputs(ext, prefix, outdir = ".")
}
\arguments{
\item{ext}{vector of file extensions}

\item{prefix}{name of file prefix for each extension.}

\item{outdir}{directory the files will be inside}
}
\value{
vector of valid file paths
}
\description{
Ext or prefix can be a vector or single character. The shorter value will be
propagated across all values of the other. See Examples for details.
}
\examples{
\dontrun{
# Checks many file types of same prefix
# ie myFile.txt, myFile.html, myFile.xml
check_outputs(c("txt", "html", "xml"), "myFile")
# Checks many files of same type
# ie myFile1.txt, myFile2.txt, myFile3.txt
check_outputs("txt", c("myFile1", "myFile2", "myFile3"))
}


}
