% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{check_valid_command_path}
\alias{check_valid_command_path}
\title{Checks path is valid}
\usage{
check_valid_command_path(path)
}
\arguments{
\item{path}{path to file or directory}
}
\value{
expanded system path
}
\description{
Not meant to be called directly
}
\examples{
\dontrun{
# will return /full/path/to/meme/bin, or error if path doesn't exist
check_valid_command_path("~/meme/bin")
}
}
