% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{check_valid_util}
\alias{check_valid_util}
\title{Checks for valid members of subdirectory}
\usage{
check_valid_util(util, utils = NULL, path = NULL)
}
\arguments{
\item{util}{name of target located in path}

\item{utils}{name of supported targets in path}

\item{path}{path to directory}
}
\value{
safe path to util, or error if util does not exist
}
\description{
Not meant to be called directly
}
\examples{
\dontrun{
# this will return /full/path/to/meme/bin/dreme
# or return an error for all values of util that are not "dreme" and "ame"
# or error if "dreme" does not exist in "~/meme/bin/"
check_valid_util("dreme", utils = c("dreme", "ame"), "~/meme/bin")

# This will throw error
check_valid_util("badUtil", utils = c("dreme", "ame"), "~/meme/bin")
}
}
