% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_args.R
\name{cmd_args_dots}
\alias{cmd_args_dots}
\title{return function dots from parent function as named list}
\usage{
cmd_args_dots(keep = NULL, drop = NULL)
}
\arguments{
\item{keep}{name of arguments to keep}

\item{drop}{name of arguments to drop (NOTE: keep or drop are mutually exclusive settings)}
}
\value{
named list of kwargs from ...
}
\description{
return function dots from parent function as named list
}
\examples{
theFunction <- function(...) { cmd_args_dots() }
theDots <-  theFunction(example = "hello", boolFlag = TRUE, vectorFlag = c(1,2,3))
}
