% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmd_drop_flags}
\alias{cmd_drop_flags}
\title{Drop entries from list of flags by name or name/value pair.}
\usage{
cmd_drop_flags(flags, drop)
}
\arguments{
\item{flags}{named list output of cmd_args_to_flags}

\item{drop}{vector of flag entries to drop. Pass an unnamed vector
to drop flags by name. Pass a named vector to drop flags by name/value
pairs. Pass a numeric vector to drop by position.}
}
\value{
flags list with values in drop removed
}
\description{
Drop entries from list of flags by name or name/value pair.
}
\examples{
exFlags <- list("flag1" = 2, "flag2" = "someText")
cmd_drop_flags(exFlags, "flag1")
# will drop flag2 because its name and value match 'drop' vector
cmd_drop_flags(exFlags, c("flag2" = "someText"))

# won't drop flag2 because its value isn't 'someText'
exFlags2 <- list("flag1" = 2, "flag2" = "otherText")
cmd_drop_flags(exFlags, c("flag2" = "someText"))
}
