% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmd_drop_list_by_name}
\alias{cmd_drop_list_by_name}
\title{Drop items by name from list}
\usage{
cmd_drop_list_by_name(list, names)
}
\arguments{
\item{list}{an R list}

\item{names}{vector of names to drop}
}
\value{
list removing items defined by names
}
\description{
Drop items by name from list
}
\examples{
cmd_drop_list_by_name(list("a" = 1, "b" = 2), "a")
}
