% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmd_error_if_missing}
\alias{cmd_error_if_missing}
\title{Check that file(s) exist, error if not}
\usage{
cmd_error_if_missing(files)
}
\arguments{
\item{files}{list or vector of paths to check}
}
\value{
nothing or error message for each missing file
}
\description{
Check that file(s) exist, error if not
}
\examples{
cmd_error_if_missing(tempdir())
\dontrun{
# Throws error if file doesn't exist
cmd_error_if_missing(file.path(tempdir(), "notreal"))
}
}
