% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmd_file_combn}
\alias{cmd_file_combn}
\title{Generates list of expected output files}
\usage{
cmd_file_combn(prefix, ext, outdir = ".")
}
\arguments{
\item{prefix}{file name to be given each ext. If a character vector, must be equal length of ext or shorter}

\item{ext}{file extension (no ".", ie "txt", "html")}

\item{outdir}{optional directory where files should exist}
}
\value{
list of file paths by each ext or prefix (whichever is longer)
}
\description{
See documentation of cmd_file_expect() for more details about how this works
}
\examples{
# Makes list for many file types of same prefix
# ie myFile.txt, myFile.html, myFile.xml
cmd_file_combn("myFile", c("txt", "html", "xml"))

# Makes list for many files of same type
# ie myFile1.txt, myFile2.txt, myFile3.txt
cmd_file_combn(c("myFile1", "myFile2", "myFile3"), "txt")

}
