% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmd_files_exist}
\alias{cmd_files_exist}
\title{Check that file exits, error if not}
\usage{
cmd_files_exist(files)
}
\arguments{
\item{files}{list or vector of paths to check}
}
\value{
nothing or error message
}
\description{
Check that file exits, error if not
}
\examples{
cmd_files_exist(tempdir())
\dontrun{
# Throws error if file doesn't exist
cmd_files_exist(file.path(tempdir(), "notreal"))
}
}
