% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_help.R
\name{cmd_help_flags_suggest}
\alias{cmd_help_flags_suggest}
\title{Error & Suggest different flag name to user}
\usage{
cmd_help_flags_suggest(suggest_names)
}
\arguments{
\item{suggest_names}{named character vector, names correspond to original
value, values correspond to suggested replacement.}
}
\value{
error message suggesting alternatives to user
}
\description{
Error & Suggest different flag name to user
}
\examples{
flags <- list("output" = "somevalue", "missplld" = "anotherValue")
suggestions <- cmd_help_flags_similar(c("output", "misspelled"), flags)
\dontrun{
cmd_error_cmd_help_flags_similar(suggestions)
}
}
