% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmd_install_check}
\alias{cmd_install_check}
\title{Wrapper function for checking an install}
\usage{
cmd_install_check(path_search, path = NULL)
}
\arguments{
\item{path_search}{\code{function} output of \code{cmd_path_search()}}

\item{path}{user-override path to check (identical to \code{path} argument of \code{cmd_path_search()} output)}
}
\value{
pretty printed message indicating whether files exits or not. Green check = Yes, red X = No.
}
\description{
This function can be lightly wrapped by package builders to build a user-friendly install checking function.
}
\examples{
\dontrun{
path_search <- cmd_path_search(default = "/bin", utils = "ls")
cmd_install_check(path_search)
}
}
