% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmd_keep_list_by_name}
\alias{cmd_keep_list_by_name}
\title{Keep items by name from list}
\usage{
cmd_keep_list_by_name(list, names)
}
\arguments{
\item{list}{an R list}

\item{names}{vector of names to keep}
}
\value{
list keeping only items defined by names
}
\description{
Keep items by name from list
}
\examples{
cmd_keep_list_by_name(list("a" = 1, "b" = 2), "a")
}
