% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmd_output_expect}
\alias{cmd_output_expect}
\title{Generates list of expected output files}
\usage{
cmd_output_expect(ext, prefix, outdir = ".")
}
\arguments{
\item{ext}{file extension (no ".", ie "txt", "html")}

\item{prefix}{file name to be given each ext. If a character vector, must be equal length of ext or shorter}

\item{outdir}{optional directory where files should exist}
}
\value{
list of file paths by each ext or prefix (whichever is longer)
}
\description{
See documentation of cmd_output_check() for more details about how this works
}
\examples{
# Makes list for many file types of same prefix
# ie myFile.txt, myFile.html, myFile.xml
cmd_output_expect(c("txt", "html", "xml"), "myFile")

# Makes list for many files of same type
# ie myFile1.txt, myFile2.txt, myFile3.txt
cmd_output_expect("txt", c("myFile1", "myFile2", "myFile3"))

}
