% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_help.R
\name{cmd_suggest_flag_names}
\alias{cmd_suggest_flag_names}
\title{Suggest alternative name by Levenshtein edit distance}
\usage{
cmd_suggest_flag_names(command_flag_names, flags, .fun = NULL)
}
\arguments{
\item{command_flag_names}{character vector of valid names (can be output of \code{\link{cmd_help_flag_names}})}

\item{flags}{a vector names correspond to values to be checked against \code{command_flag_names}}

\item{.fun}{optional function to apply to \code{command_flag_names} and \code{flags}
before checking their values. If using a function to rename flags after
\code{cmd_args_to_flags}, use that same function here. Can be useful for parsing help
lines into R-friendly variable names for user-convenience. Can be function
or \code{rlang}-style formula defintion (ie \code{.fun = ~{foo(.x)}} is the same as
\code{.fun = function(x){foo(x)}}). Note: if command_flag_names need additional
parsing after \code{\link{cmd_help_flag_names}}, it is best to do that
preprocessing before passing them to this function.}
}
\value{
named vector where names are names from \code{flags} and their values are the suggested best match from \code{command_flag_names}
}
\description{
Suggest alternative name by Levenshtein edit distance
}
\examples{
# with a flagsList, need to pass names()
flagsList <- list("output" = "somevalue", "missplld" = "anotherValue")
cmd_suggest_flag_names(c("output", "misspelled"), names(flagsList))

command_flags <- c("long-flag-name")
flags <- c("long_flag_naee")
cmd_suggest_flag_names(command_flags, flags, .fun = ~{gsub("-", "_", .x)})

# returns NULL if no errors
cmd_suggest_flag_names(c("test"), "test")
}
