% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmd_ui_file_exists}
\alias{cmd_ui_file_exists}
\title{Checks if file exists, returns pretty status message}
\usage{
cmd_ui_file_exists(file)
}
\arguments{
\item{file}{path to file}
}
\value{
ui_done or ui_oops printed to terminal.
}
\description{
Checks if file exists, returns pretty status message
}
\examples{
cmd_ui_file_exists("/path/to/file.txt")
}
