% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concatenate_args}
\alias{concatenate_args}
\title{Combines key/value pairs from named vector by separator}
\usage{
concatenate_args(dict, sep = "/")
}
\arguments{
\item{dict}{argsDict}

\item{sep}{separator}
}
\value{

}
\description{
Combines key/value pairs from named vector by separator
}
\examples{
dict <- c("name" = "val")
\dontrun{
concatenate_args(dict)
}
}
