% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_logical_to_empty}
\alias{convert_logical_to_empty}
\title{Replaces list entries of bool of specified value with empty strings}
\usage{
convert_logical_to_empty(list, bool = T)
}
\arguments{
\item{list}{a list}

\item{bool}{whether to convert T or F to empty string (default T)}
}
\value{
same list with entries replaced
}
\description{
Replaces list entries of bool of specified value with empty strings
}
\examples{
myList <- list(a = TRUE, b = FALSE)
\dontrun{
convert_logical_to_empty(myList)
}
}
