% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_names}
\alias{convert_names}
\title{converts names of object (obj) to values of named vector (dict)}
\usage{
convert_names(obj, dict)
}
\arguments{
\item{obj}{list or vector}

\item{dict}{named vector where names are converted to values}
}
\value{

}
\description{
converts names of object (obj) to values of named vector (dict)
}
\examples{
dict <- c("long" = "l")
dots <- list("long" = 1, "a" = 1)
\dontrun{
convert_names(dots, dict)
}
}
