% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_matched_args}
\alias{count_matched_args}
\title{count number of times dots contains match to long or short version of flag}
\usage{
count_matched_args(value, name, dots)
}
\arguments{
\item{value}{value from argsDict entry}

\item{name}{name of argsDict entry}

\item{dots}{list of dots from getDots()}
}
\value{
count of matches to value or name (should be 1 or greater if set, 0 if not)
}
\description{
count number of times dots contains match to long or short version of flag
}
\examples{
dict <- c("name" = "value")
dots <- list("name" = 1, "value" = 1)
\dontrun{
count_matched_args(dict[1], names(dict)[1], dots)
}

dict <- c("name" = "value")
dots <- list("name" = 1)
\dontrun{
count_matched_args(dict[1], names(dict)[1], dots)
}
}
