% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_to_args.R
\name{crystallize_flags}
\alias{crystallize_flags}
\title{Convert flag list to vector of command flags}
\usage{
crystallize_flags(flagList, prefix = "-", sep = ",")
}
\arguments{
\item{flagList}{output from argsToFlags(). A named list where names
correspond to flags and members correspond to the value for the flag.}

\item{prefix}{flag prefix, usually "-" or "--".}

\item{sep}{seperator to use if flag has a vector of values (default: NULL).}
}
\value{
character vector of parsed commandline flags followed by their values
}
\description{
Convert flag list to vector of command flags
}
\examples{
theFunction <- function(...){getAllArgs()}
theArgs <- theFunction(arg1 = "value", arg2 = TRUE)
flagList <- argsToFlags(theArgs)
flags <- crystallize_flags(flagList)

# Above is equivalent to the legacy usage:
theFunction <- function(...){getAllArgs()}
theArgs <- theFunction(arg1 = "value", arg2 = TRUE)
flags <- legacy_argsToFlags(theArgs)
}
