% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_flags_regex}
\alias{drop_flags_regex}
\title{Drop values from vector matching one or more regexes}
\usage{
drop_flags_regex(flags, regex)
}
\arguments{
\item{flags}{character vector of flags (typically output of argsToFlags())}

\item{regex}{vector of regexes to scan flags. Will remove any flags matching the regex.}
}
\value{

}
\description{
Sometimes it is easier to preprocess arguments to commandline flags before
removing known invalid entries or special circumstances. This function allows
quick removal of values by scanning the flags vector with multiple regexes.
}
\examples{
flags <- c("-n value", "-f")
drop_flags_regex(flags, "-n")
drop_flags_regex(flags, "-n value")
}
