% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_list_by_name}
\alias{drop_list_by_name}
\title{Drop items by name from list}
\usage{
drop_list_by_name(list, names)
}
\arguments{
\item{list}{an R list}

\item{names}{vector of names to drop}
}
\value{

}
\description{
Drop items by name from list
}
\examples{
\dontrun{
drop_list_by_name(list("a" = 1, "b" = 2), "a")
}
}
