% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_list_fun}
\alias{drop_list_fun}
\title{Drops list members by function}
\usage{
drop_list_fun(list, fun)
}
\arguments{
\item{list}{any list object}

\item{fun}{a function that evaluates to boolean value, used to filter members of list}
}
\value{
the same list without any entries with logical vectors
}
\description{
Drops list members by function
}
\examples{
myList <- list(a = 1, b = TRUE, c = FALSE)
\dontrun{
# This will drop logicals from list
drop_list_fun(myList, fun = is.logical)
}
}
