% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_list_logicals}
\alias{drop_list_logicals}
\title{Removes logical vectors from list objects}
\usage{
drop_list_logicals(list)
}
\arguments{
\item{list}{any list object}
}
\value{
the same list without any entries with logical vectors
}
\description{
Removes logical vectors from list objects
}
