% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_list_NULL}
\alias{drop_list_NULL}
\title{Removes logical vectors from list objects}
\usage{
drop_list_NULL(list)
}
\arguments{
\item{list}{any list object}
}
\value{
the same list without any entries with logical vectors
}
\description{
Removes logical vectors from list objects
}
\examples{
myList <- list(a = 1, b = TRUE, c = FALSE, d = NULL)
\dontrun{
drop_list_NULL(myList)
}
}
