% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_illegal_flag}
\alias{error_illegal_flag}
\title{Prints illegal flag warning for name}
\usage{
error_illegal_flag(name)
}
\arguments{
\item{name}{name containing illegal flag}
}
\value{

}
\description{
Prints illegal flag warning for name
}
\examples{
\dontrun{
error_illegal_flag("&&echo")
}
}
