% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_multimatched_args}
\alias{find_multimatched_args}
\title{returns list of named args with multiple definitions}
\usage{
find_multimatched_args(vec)
}
\arguments{
\item{vec}{named vector output of count_matched_args}
}
\value{
names of args with matches > 1
}
\description{
returns list of named args with multiple definitions
}
\examples{
vec <- c("setArg" = 1, "unsetArg" = 0, "multiSetArg" = 3)
\dontrun{
find_multimatched_args(vec)
}
}
