% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flag_is_illegal}
\alias{flag_is_illegal}
\title{Searches for illegal values in string}
\usage{
flag_is_illegal(
  flag,
  illegal_chars = c("&", "\\\\|", ";", "\\\\(", "\\\\)", "\\\\{", "\\\\}", "\\\\$",
    "\\\\@", "\\\\/", " ")
)
}
\arguments{
\item{flag}{string}

\item{illegal_chars}{vector of flags forbidden in string
default values are: "&", "|", ";", "(", ")", "{", "}", "$", "\@", "/"}
}
\value{
Boolean value for each flag
}
\description{
Searches for illegal values in string
}
\examples{
\dontrun{
flag_is_illegal("&&echo")
}
}
