% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_to_args.R
\name{flagListToFlags}
\alias{flagListToFlags}
\title{Convert flag list to vector of command flags}
\usage{
flagListToFlags(flagList, prefix = "-", sep = ",")
}
\arguments{
\item{flagList}{output from argsToFlagList(). A named list where names
correspond to flags and members correspond to the value for the flag.}

\item{prefix}{flag prefix, usually "-" or "--".}

\item{sep}{seperator to use if flag has a vector of values (default: NULL).}
}
\value{
character vector of parsed commandline flags followed by their values
}
\description{
Convert flag list to vector of command flags
}
\examples{
theFunction <- function(...){getAllArgs()}
theArgs <- theFunction(arg1 = "value", arg2 = TRUE)
flagList <- argsToFlagList(theArgs)
flags <- flagListToFlags(flagList)

# Above is equivalent to:
theFunction <- function(...){getAllArgs()}
theArgs <- theFunction(arg1 = "value", arg2 = TRUE)
flags <- argsToFlags(theArgs)
}
