% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_to_args.R
\name{getAllArgs}
\alias{getAllArgs}
\title{Return all named arguments and arguments passed as dots from parent function call}
\usage{
getAllArgs()
}
\value{
named list of all arguments passed to parent
}
\description{
Return all named arguments and arguments passed as dots from parent function call
}
\examples{
theFunction <- function(arg1, ...) { getAllArgs() }
theArgs <-  theFunction(arg1 = "test", example = "hello")
}
