% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_to_args.R
\name{getDots}
\alias{getDots}
\title{return function dots as named list}
\usage{
getDots(...)
}
\arguments{
\item{...}{dots from function call}
}
\value{
named list of kwargs from ...
}
\description{
return function dots as named list
}
\examples{
theFunction <- function(...) { getDots(...) }
theDots <-  theFunction(example = "hello", example2 = "world", boolFlag = TRUE, vectorFlag = c(1,2,3))
}
