% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_to_args.R
\name{getNamedArgs}
\alias{getNamedArgs}
\title{Return all named arguments from parent function call}
\usage{
getNamedArgs()
}
\value{
named list of all defined function arguments from parent
}
\description{
Return all named arguments from parent function call
}
\examples{
theFunction <- function(arg1, ...) { getNamedArgs() }
theNamedArgs <-  theFunction(arg1 = "test", example = "hello")
}
