% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_to_args.R
\name{legacy_argsToFlags}
\alias{legacy_argsToFlags}
\title{convert list from get*Args to vector of formatted shell flags}
\usage{
legacy_argsToFlags(args, flag_lookup = NULL, prefix = "-", sep = ",")
}
\arguments{
\item{args}{output from get*Args() family of functions (a named list)}

\item{flag_lookup}{optional named vector to convert names from dots to a
simplified shell flag, useful for providing aliases to single-letter flags}

\item{prefix}{flag prefix, usually - or --}

\item{sep}{separator to use when passing vector for a single flag. Default:
",". sep = NULL will pass each member as a new entry. Other common values could include: " ", or "\\t".}
}
\value{
named vector of shell flags followed by their values
}
\description{
convert list from get*Args to vector of formatted shell flags
}
\examples{
theFunction <- function(...) { getDotArgs() }
theDots <-  theFunction(example = "hello", boolFlag = TRUE, vectorFlag = c(1,2,3))
theFlags <-  legacy_argsToFlags(theDots)
}
