% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{warn_multimatched_arg}
\alias{warn_multimatched_arg}
\title{Prints message warning user of each argument that is defined multiple times in function call}
\usage{
warn_multimatched_arg(name)
}
\arguments{
\item{name}{name of arg}
}
\value{

}
\description{
Prints message warning user of each argument that is defined multiple times in function call
}
\examples{
\dontrun{
warn_multimatched_arg("arg")
}
}
