% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmd_list_drop_named}
\alias{cmd_list_drop_named}
\title{Drop items by name from list}
\usage{
cmd_list_drop_named(list, names)
}
\arguments{
\item{list}{an R list}

\item{names}{vector of names to drop}
}
\value{
list removing items defined by names
}
\description{
A pipe-friendly wrapper around \code{list[!(names(list) \%in\% names)]}
This function is slightly faster than using \code{\link[=cmd_list_drop]{cmd_list_drop()}} to drop items
by name.
}
\examples{
cmd_list_drop_named(list("a" = 1, "b" = 2), "a")
}
