% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_to_args.R
\name{dotsToArgs}
\alias{dotsToArgs}
\title{convert list from getDots to vector of formatted shell flags}
\usage{
dotsToArgs(dots, flag_lookup = NULL, prefix = "-")
}
\arguments{
\item{dots}{output from getDots}

\item{flag_lookup}{optional named vector to convert names from dots to a
simplified shell flag, useful for providing aliases to single-letter flags}

\item{prefix}{flag prefix, usually - or --}
}
\value{
named vector of shell flags with their values
}
\description{
convert list from getDots to vector of formatted shell flags
}
\examples{
theFunction <- function(...) { getDots(...) }
theDots <-  myFunction(example = "hello", example2 = "world", boolFlag = TRUE)
theArgs <-  dotsToArgs(theDots)
}
