% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_to_args.R
\name{getDotArgs}
\alias{getDotArgs}
\title{return function dots from parent function as named list}
\usage{
getDotArgs(keep = NULL, drop = NULL)
}
\arguments{
\item{keep}{name of arguments to keep}

\item{drop}{name of arguments to drop (NOTE: keep or drop are mutually exclusive settings)}
}
\value{
named list of kwargs from ...
}
\description{
return function dots from parent function as named list
}
\examples{
theFunction <- function(...) { getDotArgs() }
theDots <-  theFunction(example = "hello", boolFlag = TRUE, vectorFlag = c(1,2,3))
}
