% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{true_to_empty}
\alias{true_to_empty}
\title{Replace list entries with boolean values of TRUE with empty string}
\usage{
true_to_empty(list)
}
\arguments{
\item{list}{a list}
}
\value{
list where entries with TRUE are replaced for ""
}
\description{
Replace list entries with boolean values of TRUE with empty string
}
\examples{
myList <- list(a = TRUE)
\dontrun{
true_to_empty(myList)
}
}
